Refactor NodeDockerApi and NodeRpcApi
