--- it/src/test/scala/co/topl/it/IntegrationConstants.scala
@@ -1,11 +1,0 @@
-package co.topl.it
-
-import com.typesafe.config.{Config, ConfigFactory}
-
-import scala.jdk.CollectionConverters.asScalaBufferConverter
-
-trait IntegrationConstants {
-
-  val nodesJointConfig: Config = ConfigFactory.parseResources("nodes.conf").resolve()
-  val nodeSeedConfigs: List[Config] = nodesJointConfig.getConfigList("nodes").asScala.toList
-}
