+++ it/src/test/scala/co/topl/it/util/DockerSupport.scala
@@ -2,7 +2,6 @@ package co.topl.it.util
 
 import akka.actor.ActorSystem
 import co.topl.buildinfo.bifrost.BuildInfo
-import co.topl.it.api.NodeRpcApi
 import co.topl.utils.Logging
 import com.spotify.docker.client.DockerClient
 import com.spotify.docker.client.messages.ContainerConfig
